Add docs, move the view model
